third_party/blink/web_tests/virtual/popover-hint-disabled/external/wpt/html/semantics/popovers/popover-target-action-hover.tentative-expected.txt

This is a testharness.js-based test.
Found 54 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] popovertargetaction=hover shows a popover with popover=auto, invokerType=plain
  promise_test: Unhandled rejection with value: object "TypeError: Cannot read properties of undefined (reading 'slice')"
[FAIL] popovertargetaction=hover should also allow click activation, for popover=auto, invokerType=plain
  promise_test: Unhandled rejection with value: object "TypeError: Cannot read properties of undefined (reading 'slice')"
[FAIL] popovertargetaction=hover popover-show-delay is respected (popover=auto, invokerType=plain)
  promise_test: Unhandled rejection with value: object "TypeError: Cannot read properties of undefined (reading 'slice')"
[FAIL] popovertargetaction=hover does nothing when popover is already showing (popover=auto, invokerType=plain)
  promise_test: Unhandled rejection with value: object "TypeError: Cannot read properties of undefined (reading 'slice')"
[FAIL] popovertargetaction=hover does nothing when popover is moved out of the document (popover=auto, invokerType=plain)
  promise_test: Unhandled rejection with value: object "TypeError: Cannot read properties of undefined (reading 'slice')"
[FAIL] popovertargetaction=hover does nothing when target changes (popover=auto, invokerType=plain)
  promise_test: Unhandled rejection with value: object "TypeError: Cannot read properties of undefined (reading 'slice')"
[FAIL] popovertargetaction=hover shows a popover with popover=auto, invokerType=nested
  promise_test: Unhandled rejection with value: object "TypeError: Cannot read properties of undefined (reading 'slice')"
[FAIL] popovertargetaction=hover should also allow click activation, for popover=auto, invokerType=nested
  promise_test: Unhandled rejection with value: object "TypeError: Cannot read properties of undefined (reading 'slice')"
[FAIL] popovertargetaction=hover popover-show-delay is respected (popover=auto, invokerType=nested)
  promise_test: Unhandled rejection with value: object "TypeError: Cannot read properties of undefined (reading 'slice')"
[FAIL] popovertargetaction=hover does nothing when popover is already showing (popover=auto, invokerType=nested)
  promise_test: Unhandled rejection with value: object "TypeError: Cannot read properties of undefined (reading 'slice')"
[FAIL] popovertargetaction=hover does nothing when popover is moved out of the document (popover=auto, invokerType=nested)
  promise_test: Unhandled rejection with value: object "TypeError: Cannot read properties of undefined (reading 'slice')"
[FAIL] popovertargetaction=hover does nothing when target changes (popover=auto, invokerType=nested)
  promise_test: Unhandled rejection with value: object "TypeError: Cannot read properties of undefined (reading 'slice')"
[FAIL] popovertargetaction=hover shows a popover with popover=auto, invokerType=nested-offset
  promise_test: Unhandled rejection with value: object "TypeError: Cannot read properties of undefined (reading 'slice')"
[FAIL] popovertargetaction=hover should also allow click activation, for popover=auto, invokerType=nested-offset
  promise_test: Unhandled rejection with value: object "TypeError: Cannot read properties of undefined (reading 'slice')"
[FAIL] popovertargetaction=hover popover-show-delay is respected (popover=auto, invokerType=nested-offset)
  promise_test: Unhandled rejection with value: object "TypeError: Cannot read properties of undefined (reading 'slice')"
[FAIL] popovertargetaction=hover does nothing when popover is already showing (popover=auto, invokerType=nested-offset)
  promise_test: Unhandled rejection with value: object "TypeError: Cannot read properties of undefined (reading 'slice')"
[FAIL] popovertargetaction=hover does nothing when popover is moved out of the document (popover=auto, invokerType=nested-offset)
  promise_test: Unhandled rejection with value: object "TypeError: Cannot read properties of undefined (reading 'slice')"
[FAIL] popovertargetaction=hover does nothing when target changes (popover=auto, invokerType=nested-offset)
  promise_test: Unhandled rejection with value: object "TypeError: Cannot read properties of undefined (reading 'slice')"
[FAIL] popovertargetaction=hover shows a popover with popover=hint, invokerType=plain
  promise_test: Unhandled rejection with value: object "TypeError: Cannot read properties of undefined (reading 'slice')"
[FAIL] popovertargetaction=hover should also allow click activation, for popover=hint, invokerType=plain
  promise_test: Unhandled rejection with value: object "TypeError: Cannot read properties of undefined (reading 'slice')"
[FAIL] popovertargetaction=hover popover-show-delay is respected (popover=hint, invokerType=plain)
  promise_test: Unhandled rejection with value: object "TypeError: Cannot read properties of undefined (reading 'slice')"
[FAIL] popovertargetaction=hover does nothing when popover is already showing (popover=hint, invokerType=plain)
  promise_test: Unhandled rejection with value: object "TypeError: Cannot read properties of undefined (reading 'slice')"
[FAIL] popovertargetaction=hover does nothing when popover is moved out of the document (popover=hint, invokerType=plain)
  promise_test: Unhandled rejection with value: object "TypeError: Cannot read properties of undefined (reading 'slice')"
[FAIL] popovertargetaction=hover does nothing when target changes (popover=hint, invokerType=plain)
  promise_test: Unhandled rejection with value: object "TypeError: Cannot read properties of undefined (reading 'slice')"
[FAIL] popovertargetaction=hover shows a popover with popover=hint, invokerType=nested
  promise_test: Unhandled rejection with value: object "TypeError: Cannot read properties of undefined (reading 'slice')"
[FAIL] popovertargetaction=hover should also allow click activation, for popover=hint, invokerType=nested
  promise_test: Unhandled rejection with value: object "TypeError: Cannot read properties of undefined (reading 'slice')"
[FAIL] popovertargetaction=hover popover-show-delay is respected (popover=hint, invokerType=nested)
  promise_test: Unhandled rejection with value: object "TypeError: Cannot read properties of undefined (reading 'slice')"
[FAIL] popovertargetaction=hover does nothing when popover is already showing (popover=hint, invokerType=nested)
  promise_test: Unhandled rejection with value: object "TypeError: Cannot read properties of undefined (reading 'slice')"
[FAIL] popovertargetaction=hover does nothing when popover is moved out of the document (popover=hint, invokerType=nested)
  promise_test: Unhandled rejection with value: object "TypeError: Cannot read properties of undefined (reading 'slice')"
[FAIL] popovertargetaction=hover does nothing when target changes (popover=hint, invokerType=nested)
  promise_test: Unhandled rejection with value: object "TypeError: Cannot read properties of undefined (reading 'slice')"
[FAIL] popovertargetaction=hover shows a popover with popover=hint, invokerType=nested-offset
  promise_test: Unhandled rejection with value: object "TypeError: Cannot read properties of undefined (reading 'slice')"
[FAIL] popovertargetaction=hover should also allow click activation, for popover=hint, invokerType=nested-offset
  promise_test: Unhandled rejection with value: object "TypeError: Cannot read properties of undefined (reading 'slice')"
[FAIL] popovertargetaction=hover popover-show-delay is respected (popover=hint, invokerType=nested-offset)
  promise_test: Unhandled rejection with value: object "TypeError: Cannot read properties of undefined (reading 'slice')"
[FAIL] popovertargetaction=hover does nothing when popover is already showing (popover=hint, invokerType=nested-offset)
  promise_test: Unhandled rejection with value: object "TypeError: Cannot read properties of undefined (reading 'slice')"
[FAIL] popovertargetaction=hover does nothing when popover is moved out of the document (popover=hint, invokerType=nested-offset)
  promise_test: Unhandled rejection with value: object "TypeError: Cannot read properties of undefined (reading 'slice')"
[FAIL] popovertargetaction=hover does nothing when target changes (popover=hint, invokerType=nested-offset)
  promise_test: Unhandled rejection with value: object "TypeError: Cannot read properties of undefined (reading 'slice')"
[FAIL] popovertargetaction=hover shows a popover with popover=manual, invokerType=plain
  promise_test: Unhandled rejection with value: object "TypeError: Cannot read properties of undefined (reading 'slice')"
[FAIL] popovertargetaction=hover should also allow click activation, for popover=manual, invokerType=plain
  promise_test: Unhandled rejection with value: object "TypeError: Cannot read properties of undefined (reading 'slice')"
[FAIL] popovertargetaction=hover popover-show-delay is respected (popover=manual, invokerType=plain)
  promise_test: Unhandled rejection with value: object "TypeError: Cannot read properties of undefined (reading 'slice')"
[FAIL] popovertargetaction=hover does nothing when popover is already showing (popover=manual, invokerType=plain)
  promise_test: Unhandled rejection with value: object "TypeError: Cannot read properties of undefined (reading 'slice')"
[FAIL] popovertargetaction=hover does nothing when popover is moved out of the document (popover=manual, invokerType=plain)
  promise_test: Unhandled rejection with value: object "TypeError: Cannot read properties of undefined (reading 'slice')"
[FAIL] popovertargetaction=hover does nothing when target changes (popover=manual, invokerType=plain)
  promise_test: Unhandled rejection with value: object "TypeError: Cannot read properties of undefined (reading 'slice')"
[FAIL] popovertargetaction=hover shows a popover with popover=manual, invokerType=nested
  promise_test: Unhandled rejection with value: object "TypeError: Cannot read properties of undefined (reading 'slice')"
[FAIL] popovertargetaction=hover should also allow click activation, for popover=manual, invokerType=nested
  promise_test: Unhandled rejection with value: object "TypeError: Cannot read properties of undefined (reading 'slice')"
[FAIL] popovertargetaction=hover popover-show-delay is respected (popover=manual, invokerType=nested)
  promise_test: Unhandled rejection with value: object "TypeError: Cannot read properties of undefined (reading 'slice')"
[FAIL] popovertargetaction=hover does nothing when popover is already showing (popover=manual, invokerType=nested)
  promise_test: Unhandled rejection with value: object "TypeError: Cannot read properties of undefined (reading 'slice')"
[FAIL] popovertargetaction=hover does nothing when popover is moved out of the document (popover=manual, invokerType=nested)
  promise_test: Unhandled rejection with value: object "TypeError: Cannot read properties of undefined (reading 'slice')"
[FAIL] popovertargetaction=hover does nothing when target changes (popover=manual, invokerType=nested)
  promise_test: Unhandled rejection with value: object "TypeError: Cannot read properties of undefined (reading 'slice')"
[FAIL] popovertargetaction=hover shows a popover with popover=manual, invokerType=nested-offset
  promise_test: Unhandled rejection with value: object "TypeError: Cannot read properties of undefined (reading 'slice')"
[FAIL] popovertargetaction=hover should also allow click activation, for popover=manual, invokerType=nested-offset
  promise_test: Unhandled rejection with value: object "TypeError: Cannot read properties of undefined (reading 'slice')"
[FAIL] popovertargetaction=hover popover-show-delay is respected (popover=manual, invokerType=nested-offset)
  promise_test: Unhandled rejection with value: object "TypeError: Cannot read properties of undefined (reading 'slice')"
[FAIL] popovertargetaction=hover does nothing when popover is already showing (popover=manual, invokerType=nested-offset)
  promise_test: Unhandled rejection with value: object "TypeError: Cannot read properties of undefined (reading 'slice')"
[FAIL] popovertargetaction=hover does nothing when popover is moved out of the document (popover=manual, invokerType=nested-offset)
  promise_test: Unhandled rejection with value: object "TypeError: Cannot read properties of undefined (reading 'slice')"
[FAIL] popovertargetaction=hover does nothing when target changes (popover=manual, invokerType=nested-offset)
  promise_test: Unhandled rejection with value: object "TypeError: Cannot read properties of undefined (reading 'slice')"
Harness: the test ran to completion.
